Source updated to 38.0. Compilation with ICL happens without any issues at all during compile time! Big improvement from last commits. You should have no issues compiling Waterfox

+++ browser/components/loop/test/functional/serversetup.py
@@ -23,8 +23,8 @@ CONTENT_SERVER_ENV = os.environ.copy()
 CONTENT_SERVER_ENV.update({"PORT": str(CONTENT_SERVER_PORT),
                            "LOOP_SERVER_PORT": str(LOOP_SERVER_PORT)})
 
-WEB_APP_URL = "http://localhost:" + str(CONTENT_SERVER_PORT) + \
-              "/content/#call/{token}"
+ROOMS_WEB_APP_URL = "http://localhost:" + str(CONTENT_SERVER_PORT) + \
+  "/content/{token}"
 
 LOOP_SERVER_COMMAND = ["make", "runserver"]
 LOOP_SERVER_ENV = os.environ.copy()
@@ -33,17 +33,21 @@ LOOP_SERVER_ENV = os.environ.copy()
 LOOP_SERVER_ENV.update({"NODE_ENV": "dev",
                         "PORT": str(LOOP_SERVER_PORT),
                         "SERVER_ADDRESS": "localhost:" + str(LOOP_SERVER_PORT),
-                        "WEB_APP_URL": WEB_APP_URL})
+                        "ROOMS_WEB_APP_URL": ROOMS_WEB_APP_URL})
 
 
 class LoopTestServers:
     def __init__(self):
-        self.loop_server = self.start_loop_server()
+        loop_server_location = os.environ.get('LOOP_SERVER')
+        if loop_server_location.startswith("http"):
+            FIREFOX_PREFERENCES["loop.server"] = loop_server_location
+            return
+
+        self.loop_server = self.start_loop_server(loop_server_location)
         self.content_server = self.start_content_server()
 
     @staticmethod
-    def start_loop_server():
-        loop_server_location = os.environ.get('LOOP_SERVER')
+    def start_loop_server(loop_server_location):
         if loop_server_location is None:
             raise Exception('LOOP_SERVER variable not set')
 
@@ -68,6 +72,8 @@ class LoopTestServers:
         return p
 
     def shutdown(self):
-        self.content_server.kill()
-        self.loop_server.kill()
+        if hasattr(self, "content_server"):
+            self.content_server.kill()
+        if hasattr(self, "loop_server"):
+            self.loop_server.kill()
